refactor(#625): scatter VFS hooks to service dirs, extract contracts
Move hook contracts and implementations to their proper architectural layers:

- Extract OperationWarning to contracts/types.py (was in core/)
- Create contracts/vfs_hooks.py with hook context dataclasses + protocols
- Scatter 3 hook impls from core/vfs_hook_impls.py to their home services:
  - DynamicViewerReadHook → services/rebac/
  - AutoParseWriteHook → parsers/
  - TigerCacheRenameHook → services/permissions/cache/tiger/
- Move hook registration wiring from NexusFS.__init__ to factory/orchestrator.py
- Fix TigerCache double execution bug in rename() (nexus_fs_core.py)
- Add kernel notification dispatch architecture to KERNEL-ARCHITECTURE.md
- Split tests to follow their hooks (tests/unit/services/, tests/unit/parsers/)
