+++ src/nexus/contracts/types.py
@@ -19,7 +19,7 @@ import uuid
 from collections.abc import Callable
 from dataclasses import asdict, dataclass, field
 from enum import IntFlag, StrEnum
-from typing import TYPE_CHECKING, Any, Protocol, runtime_checkable
+from typing import TYPE_CHECKING, Any, Literal, Protocol, runtime_checkable
 
 if TYPE_CHECKING:
     from nexus.core.read_set import ReadSet
@@ -345,3 +345,26 @@ class SnapshotId:
     """Opaque identifier for a transactional snapshot."""
 
     id: str
+
+
+# ---------------------------------------------------------------------------
+# Operation warnings (moved from core/operation_result.py, Issue #625)
+# ---------------------------------------------------------------------------
+
+
+@dataclass(frozen=True, slots=True)
+class OperationWarning:
+    """A non-fatal warning produced during a VFS operation.
+
+    Attributes:
+        severity: "degraded" means a side-effect (cache, ReBAC, observer) failed
+                  but the core operation succeeded.
+                  "cosmetic" means something minor was skipped (e.g., logging).
+        component: Which subsystem produced the warning (e.g., "tiger_cache",
+                   "rebac", "observer", "parser").
+        message: Human-readable description of what failed.
+    """
+
+    severity: Literal["degraded", "cosmetic"]
+    component: str
+    message: str
