+++ src/nexus/core/nexus_fs_core.py
@@ -1038,7 +1038,7 @@ class NexusFSCoreMixin:
         # Issue #2033 Phase 5: Dispatch post-read hooks (dynamic viewer, tracking, etc.)
         _pipeline = getattr(self, "_hook_pipeline", None)
         if _pipeline is not None and _pipeline.read_hook_count > 0:
-            from nexus.core.vfs_hooks import ReadHookContext
+            from nexus.contracts.vfs_hooks import ReadHookContext
 
             _read_ctx = ReadHookContext(
                 path=path,
@@ -2101,7 +2101,7 @@ class NexusFSCoreMixin:
         # Issue #2033 Phase 5: Dispatch post-write hooks (auto-parse, tiger cache, etc.)
         _pipeline = getattr(self, "_hook_pipeline", None)
         if _pipeline is not None and _pipeline.write_hook_count > 0:
-            from nexus.core.vfs_hooks import WriteHookContext
+            from nexus.contracts.vfs_hooks import WriteHookContext
 
             _write_ctx = WriteHookContext(
                 path=path,
@@ -3125,7 +3125,7 @@ class NexusFSCoreMixin:
         # Issue #2033 Phase 5: Dispatch post-delete hooks
         _pipeline = getattr(self, "_hook_pipeline", None)
         if _pipeline is not None and _pipeline.delete_hook_count > 0:
-            from nexus.core.vfs_hooks import DeleteHookContext
+            from nexus.contracts.vfs_hooks import DeleteHookContext
 
             _delete_ctx = DeleteHookContext(
                 path=path,
@@ -3332,28 +3332,10 @@ class NexusFSCoreMixin:
             logger.warning("[RENAME-REBAC] SKIPPED - no _rebac_manager available")
 
         # Leopard-style: Update Tiger Cache bitmaps for moved files
-        # When a file moves between directories, permissions may change:
-        # - If old directory had grants, file should be removed from those users' bitmaps
-        # - If new directory has grants, file should be added to those users' bitmaps
-        if hasattr(self, "_rebac_manager") and self._rebac_manager:
-            try:
-                tiger_cache = getattr(self._rebac_manager, "_tiger_cache", None)
-                if tiger_cache:
-                    self._update_tiger_cache_on_move(
-                        tiger_cache=tiger_cache,
-                        old_path=old_path,
-                        new_path=new_path,
-                        is_directory=bool(is_directory),
-                        zone_id=zone_id or "root",
-                    )
-            except Exception as e:
-                # Log but don't fail the rename operation
-                logger.warning(f"[LEOPARD] Failed to update Tiger Cache on move: {e}")
-
         # Issue #2033 Phase 5: Dispatch post-rename hooks (tiger cache, etc.)
         _pipeline = getattr(self, "_hook_pipeline", None)
         if _pipeline is not None and _pipeline.rename_hook_count > 0:
-            from nexus.core.vfs_hooks import RenameHookContext
+            from nexus.contracts.vfs_hooks import RenameHookContext
 
             _rename_ctx = RenameHookContext(
                 old_path=old_path,
@@ -3407,171 +3389,6 @@ class NexusFSCoreMixin:
         )
 
         return {}
-
-    def _update_tiger_cache_on_move(
-        self,
-        tiger_cache: Any,
-        old_path: str,
-        new_path: str,
-        is_directory: bool,
-        zone_id: str,
-    ) -> None:
-        """Update Tiger Cache bitmaps when a file/directory is moved.
-
-        Leopard-style optimization: When a file moves between directories,
-        permissions may change based on ancestor directory grants:
-        - If old directory had grants, file should be removed from those users' bitmaps
-        - If new directory has grants, file should be added to those users' bitmaps
-        - Grants in both paths need no change (permission still applies)
-
-        Args:
-            tiger_cache: The TigerCache instance
-            old_path: Original file path
-            new_path: New file path
-            is_directory: Whether this is a directory move
-            zone_id: Zone ID
-        """
-        # Get grants that apply to old and new paths
-        old_grants = tiger_cache.get_directory_grants_for_path(old_path, zone_id)
-        new_grants = tiger_cache.get_directory_grants_for_path(new_path, zone_id)
-
-        # Create grant keys for comparison (subject_type, subject_id, permission)
-        def grant_key(g: dict) -> tuple:
-            return (g["subject_type"], g["subject_id"], g["permission"])
-
-        old_grant_keys = {grant_key(g) for g in old_grants}
-        new_grant_keys = {grant_key(g) for g in new_grants}
-
-        # Grants only in old path -> remove file from those bitmaps
-        grants_to_remove = old_grant_keys - new_grant_keys
-        # Grants only in new path -> add file to those bitmaps
-        grants_to_add = new_grant_keys - old_grant_keys
-
-        if not grants_to_remove and not grants_to_add:
-            logger.debug(
-                f"[LEOPARD] No permission changes needed for move: {old_path} -> {new_path}"
-            )
-            return
-
-        # Get files to update (single file or all descendants for directory)
-        if is_directory:
-            files_to_update = self._get_directory_files_for_move(old_path, new_path, zone_id)
-        else:
-            files_to_update = [(old_path, new_path)]
-
-        logger.info(
-            f"[LEOPARD] Updating permissions for move: {old_path} -> {new_path}, "
-            f"files={len(files_to_update)}, grants_to_remove={len(grants_to_remove)}, "
-            f"grants_to_add={len(grants_to_add)}"
-        )
-
-        # Process each file
-        resource_map = getattr(tiger_cache, "_resource_map", None)
-        if not resource_map:
-            logger.warning("[LEOPARD] No resource map available, skipping bitmap updates")
-            return
-
-        for _old_file_path, new_file_path in files_to_update:
-            # Get int_id for the file (use new path since file was already renamed)
-            int_id = resource_map.get_or_create_int_id("file", new_file_path)
-            if int_id <= 0:
-                logger.warning(f"[LEOPARD] Failed to get int_id for: {new_file_path}")
-                continue
-
-            # Remove from old grants' bitmaps
-            for subject_type, subject_id, permission in grants_to_remove:
-                try:
-                    tiger_cache.remove_from_bitmap(
-                        subject_type=subject_type,
-                        subject_id=subject_id,
-                        permission=permission,
-                        resource_type="file",
-                        zone_id=zone_id,
-                        resource_int_id=int_id,
-                    )
-                    logger.debug(
-                        f"[LEOPARD] Removed {new_file_path} from bitmap: "
-                        f"{subject_type}:{subject_id} ({permission})"
-                    )
-                except Exception as e:
-                    logger.warning(f"[LEOPARD] Failed to remove from bitmap: {e}")
-
-            # Add to new grants' bitmaps
-            for grant in new_grants:
-                key = grant_key(grant)
-                if key not in grants_to_add:
-                    continue
-
-                # Check if grant includes future files (moved files count as "new" to this path)
-                if not grant.get("include_future_files", True):
-                    continue
-
-                try:
-                    tiger_cache.add_to_bitmap(
-                        grant["subject_type"],
-                        grant["subject_id"],
-                        grant["permission"],
-                        "file",
-                        zone_id,
-                        int_id,
-                    )
-
-                    # Persist immediately (write-through)
-                    tiger_cache.persist_single_grant(
-                        grant["subject_type"],
-                        grant["subject_id"],
-                        grant["permission"],
-                        "file",
-                        new_file_path,
-                        zone_id,
-                    )
-
-                    logger.debug(
-                        f"[LEOPARD] Added {new_file_path} to bitmap: "
-                        f"{grant['subject_type']}:{grant['subject_id']} ({grant['permission']})"
-                    )
-                except Exception as e:
-                    logger.warning(f"[LEOPARD] Failed to add to bitmap: {e}")
-
-        logger.info(f"[LEOPARD] Completed permission updates for move: {old_path} -> {new_path}")
-
-    def _get_directory_files_for_move(
-        self,
-        old_dir_path: str,
-        new_dir_path: str,
-        zone_id: str,
-    ) -> list[tuple[str, str]]:
-        """Get all files under a directory for move permission updates.
-
-        Args:
-            old_dir_path: Original directory path
-            new_dir_path: New directory path
-            zone_id: Zone ID
-
-        Returns:
-            List of (old_file_path, new_file_path) tuples
-        """
-        # Normalize paths
-        old_prefix = old_dir_path.rstrip("/") + "/"
-        new_prefix = new_dir_path.rstrip("/") + "/"
-
-        try:
-            # Query all files under directory (using new path since already renamed)
-            # The files have already been renamed via metadata update, so query new paths
-            result = []
-            for file_meta in self.metadata.list_iter(
-                prefix=new_prefix, recursive=True, zone_id=zone_id
-            ):
-                new_file_path = file_meta.path
-                if new_file_path:
-                    # Compute what the old path would have been
-                    relative_path = new_file_path[len(new_prefix) :]
-                    old_file_path = old_prefix + relative_path
-                    result.append((old_file_path, new_file_path))
-            return result
-        except Exception as e:
-            logger.warning(f"[LEOPARD] Failed to list directory files: {e}")
-            return []
 
     @rpc_expose(description="Get file metadata without reading content")
     def stat(self, path: str, context: OperationContext | None = None) -> dict[str, Any]:
