+++ src/nexus/core/operation_result.py
@@ -23,25 +23,9 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import Generic, Literal, TypeVar
 
+from nexus.contracts.types import OperationWarning
+
 T = TypeVar("T")
-
-
-@dataclass(frozen=True, slots=True)
-class OperationWarning:
-    """A non-fatal warning produced during a VFS operation.
-
-    Attributes:
-        severity: "degraded" means a side-effect (cache, ReBAC, observer) failed
-                  but the core operation succeeded.
-                  "cosmetic" means something minor was skipped (e.g., logging).
-        component: Which subsystem produced the warning (e.g., "tiger_cache",
-                   "rebac", "observer", "parser").
-        message: Human-readable description of what failed.
-    """
-
-    severity: Literal["degraded", "cosmetic"]
-    component: str
-    message: str
 
 
 @dataclass(frozen=True, slots=True)
